server/requirements.txt
amqp==5.3.1
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
billiard==4.2.1
binaryornot==0.4.4
bleach==6.2.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
#click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
cookiecutter==2.6.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Django
django-channels==0.7.0
django-cors-headers==4.7.0
django-filter==25.1
django-forms-bootstrap==3.1.0
django-mysql==4.17.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.11
google-auth
msgpack
lz4
